requirements.txt
flask
flask_cors
flask_jwt_extended
flask_monitoringdashboard
werkzeug
python-dotenv